Refactor classic cards presenter
