+++ src/ui/views/classic/index.js
@@ -1,13 +1,15 @@
 import resolvers, { classicResolvers } from './resolvers.js';
 import { renderDashboard as baseRenderDashboard } from '../../dashboard.js';
 import dashboardPresenter from './dashboardPresenter.js';
+import cardsPresenter from './cardsPresenter.js';
 
 const classicView = {
   id: 'classic',
   name: 'Classic Dashboard',
   resolvers,
   presenters: {
-    dashboard: dashboardPresenter
+    dashboard: dashboardPresenter,
+    cards: cardsPresenter
   },
   renderDashboard(state, summary) {
     baseRenderDashboard(state, summary, dashboardPresenter);
